+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Make cleanup of files Windows compatible
 - Filenames can be paths (`foo/bar/my_pdf`).
+- Replace `filename` by `filepath` in the names of the arguments.
 
 
 ## [0.7.1]